Deleting a user from the system (correction)

+++ src/main/java/com/ecommerce/ecommerce/repository/UserRepositoryImpl.java
@@ -14,6 +14,11 @@ import java.util.List;
 @Repository
 public class UserRepositoryImpl implements UserRepository {
     private static final String USER_TABLE_NAME = "custom_user";
+    private static final String ORDER_TABLE_NAME = "orders";
+    private static final String FAVORITE_ITEM_TABLE_NAME = "favorite_item";
+    private static final String ORDER_ITEM_ITEM_TABLE_NAME = "order_item";
+
+
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
@@ -43,22 +48,25 @@ public class UserRepositoryImpl implements UserRepository {
 
     @Override
     public void deleteCustomUserById(Long id) {
-        // Delete order items associated with the user
-        String deleteOrderItemsSql = "DELETE FROM order_item " +
-                "WHERE order_id IN (SELECT id FROM orders WHERE user_id = ?)";
-        jdbcTemplate.update(deleteOrderItemsSql, id);
+        try {
+            // Delete order items associated with the user
+            String deleteOrderItemsSql = "DELETE FROM order_item WHERE order_id IN (SELECT id FROM orders WHERE user_id=?)";
+            jdbcTemplate.update(deleteOrderItemsSql, id);
 
-        // Delete orders associated with the user
-        String deleteOrdersSql = "DELETE FROM orders WHERE user_id = ?";
-        jdbcTemplate.update(deleteOrdersSql, id);
+            // Delete orders associated with the user
+            String deleteOrdersSql = "DELETE FROM " + ORDER_TABLE_NAME + " WHERE user_id=?";
+            jdbcTemplate.update(deleteOrdersSql, id);
 
-        // Delete favorite items associated with the user
-        String deleteFavoriteItemsSql = "DELETE FROM favorite_item WHERE user_id = ?";
-        jdbcTemplate.update(deleteFavoriteItemsSql, id);
+            // Delete favorite items associated with the user
+            String deleteFavoriteItemsSql = "DELETE FROM " + FAVORITE_ITEM_TABLE_NAME + " WHERE user_id=?";
+            jdbcTemplate.update(deleteFavoriteItemsSql, id);
 
-        // Delete the user
-        String deleteUserSql = "DELETE FROM " + USER_TABLE_NAME + " WHERE id=?";
-        jdbcTemplate.update(deleteUserSql, id);
+            // Delete the user
+            String deleteUserSql = "DELETE FROM " + USER_TABLE_NAME + " WHERE id=?";
+            jdbcTemplate.update(deleteUserSql, id);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
     }
 
     @Override
